UHF-6839: Move ALL query logic inside single SWR hook to prevent useless re-rendering
