+++ src/js/react/apps/school-search/hooks/UseQuery.ts
@@ -1,72 +1,57 @@
-import BooleanQuery from '@/types/BooleanQuery';
-import GlobalSettings from '../enum/GlobalSettings';
-import UseAddressQuery from './UseAddressQuery';
-import UseCoordinates from './UseCoordinates';
+import useSWR from 'swr';
+import { useAtomValue } from 'jotai';
 import SearchParams from '../types/SearchParams';
+import configurationsAtom from '../store';
+import { getCoordsUrl, getLocationsUrl, parseCoordinates } from '../helpers/SubQueries';
+import getQueryString from '../helpers/ElasticQuery';
+import GlobalSettings from '../enum/GlobalSettings';
 
 const UseQuery = (params: SearchParams) => {
+  const { baseUrl } = useAtomValue(configurationsAtom);
   const { address } = params;
-  const { size } = GlobalSettings;
   const page = Number.isNaN(Number(params.page)) ? 1 : Number(params.page);
 
-  const { coordinates, isLoading: coordsLoading, isValidating: coordsValidating, noResults } = UseCoordinates(address);
-  const { ids, isLoading: idsLoading, isValidating: idsValidating } = UseAddressQuery(coordinates);
+  const fetcher = async() => {
+    const { index } = GlobalSettings;
 
-  console.log(coordinates, ids);
+    let coordinates = null;
+    let ids = null;
 
-  const query: BooleanQuery = {
-    bool: {
-      filter: [
-        {
-          term: {
-            _language: drupalSettings.path.currentLanguage
-          }
-        }
-      ],
+    if (address) {
+      const coordinatesRes = await fetch(getCoordsUrl(address));
+      const coordinatesData = await coordinatesRes.json();
+
+      coordinates = parseCoordinates(coordinatesData);
     }
+
+    if (coordinates && coordinates.length) {
+      const [lat, lon] = coordinates;
+      const locationsResponse = await fetch(getLocationsUrl(lat, lon));
+      const locationsData = await locationsResponse.json();
+
+      if (!locationsData || !locationsData.results) {
+        return null;
+      }
+
+      ids = locationsData.results.map((result: {service_point_id: string}) => result.service_point_id);
+    }
+
+    return fetch(`${baseUrl}/${index}/_search`, {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      body: getQueryString(ids, coordinates, page),
+    }).then((res) => res.json());
   };
 
-  if (ids && ids.length) {
-    query.bool.must = [
-      {
-        terms: {
-          id: ids
-        }
-      }
-    ];
-  }
-
-  let sort: any = [
-    {
-      name: 'asc'
-    }
-  ];
-
-  if (coordinates && coordinates.length) {
-    const [lat, lon] = coordinates;
-
-   sort =[{
-      _geo_distance: {
-        coordinates: {
-          lat,
-          lon,
-        },
-        order: 'asc',
-        unit: 'm',
-        distance_type: 'arc',
-        ignore_unmapped: true
-      }
-    }, ...sort];
-  }
+  const { data, error } = useSWR(address, fetcher, {
+    revalidateOnFocus: false
+  });
 
   return {
-    noResults,
-    url: JSON.stringify({
-      from: size * (page - 1),
-      query,
-      size,
-      sort
-    })
+    data,
+    error
   };
 };
